pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chartproject"
version = "0.1.0"
description = "Reusable data pipeline foundation for market + social visualization projects"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "duckdb>=1.1.0",
  "pandas>=2.2.0",
  "pyarrow>=15.0.0",
  "python-dotenv>=1.0.0",
  "pydantic>=2.6.0",
  "plotly>=5.20.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.4.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
